add tiktok part 1

+++ controllers/profile.controller.js
@@ -8,6 +8,7 @@ const {
     Interests,
     Notification,
     FbPage,
+    TikTokProfile
 } = require('../model/index')
 
 const { responseHandler } = require('../helpers/response-handler')
@@ -98,7 +99,7 @@ module.exports.uploadUserLegal = multer({ storage: storageUserLegal }).single(
     'file'
 )
 
-exports.account = async (req, res) => {
+exports.account = async(req, res) => {
     try {
         if (req.user) {
             let { password, ...user } = req.user.toObject()
@@ -115,7 +116,7 @@ exports.account = async (req, res) => {
     }
 }
 
-exports.profilePicture = async (req, response) => {
+exports.profilePicture = async(req, response) => {
     try {
         const idUser = req.query.id ? +req.query.id : req.user._id
         gfsprofilePic.files.findOne({ 'user.$id': idUser }, (err, file) => {
@@ -140,24 +141,21 @@ exports.profilePicture = async (req, response) => {
     }
 }
 
-module.exports.addProfilePicture = async (req, res) => {
+module.exports.addProfilePicture = async(req, res) => {
     try {
         if (req.file) {
             await gfsprofilePic.files.findOneAndDelete({
                 'user.$id': req.user._id,
             })
-            await gfsprofilePic.files.updateOne(
-                { _id: req.file.id },
-                {
-                    $set: {
-                        user: {
-                            $ref: 'user',
-                            $id: req.user._id,
-                            $db: 'nodesatt',
-                        },
+            await gfsprofilePic.files.updateOne({ _id: req.file.id }, {
+                $set: {
+                    user: {
+                        $ref: 'user',
+                        $id: req.user._id,
+                        $db: 'nodesatt',
                     },
-                }
-            )
+                },
+            })
             return makeResponseData(res, 201, 'Saved')
         }
         return makeResponseData(res, 204, 'Only images allowed')
@@ -170,7 +168,7 @@ module.exports.addProfilePicture = async (req, res) => {
     }
 }
 
-exports.updateProfile = async (req, res) => {
+exports.updateProfile = async(req, res) => {
     try {
         const id = req.user._id
         let profile = req.body
@@ -183,11 +181,7 @@ exports.updateProfile = async (req, res) => {
                 return makeResponseError(res, 406, 'email already exists')
             }
         }
-        const updatedProfile = await User.findOneAndUpdate(
-            { _id: id },
-            { $set: profile },
-            { new: true }
-        ).select('-password')
+        const updatedProfile = await User.findOneAndUpdate({ _id: id }, { $set: profile }, { new: true }).select('-password')
 
         if (updatedProfile.nModified === 0) {
             return makeResponseError(res, 400, 'update failed')
@@ -203,7 +197,7 @@ exports.updateProfile = async (req, res) => {
     }
 }
 
-exports.UserLegalProfile = async (req, res) => {
+exports.UserLegalProfile = async(req, res) => {
     try {
         const idNode = '0' + req.user._id
         const files = await gfsUserLegal.files.find({ idNode }).toArray()
@@ -224,7 +218,7 @@ exports.UserLegalProfile = async (req, res) => {
     }
 }
 
-exports.addUserLegalProfile = async (req, res) => {
+exports.addUserLegalProfile = async(req, res) => {
     try {
         const id = req.user._id
         const idNode = '0' + id
@@ -234,24 +228,20 @@ exports.addUserLegalProfile = async (req, res) => {
             await gfsUserLegal.files.deleteMany({
                 $and: [{ idNode }, { type }],
             })
-            const updatedLegalProfile = await gfsUserLegal.files.updateMany(
-                { _id: req.file.id },
-                {
-                    $set: {
-                        idNode,
-                        DataUser: {
-                            $ref: 'user',
-                            $id: Long.fromNumber(id),
-                            $db: 'nodesatt',
-                        },
-                        validate: false,
-                        type,
+            const updatedLegalProfile = await gfsUserLegal.files.updateMany({ _id: req.file.id }, {
+                $set: {
+                    idNode,
+                    DataUser: {
+                        $ref: 'user',
+                        $id: Long.fromNumber(id),
+                        $db: 'nodesatt',
                     },
+                    validate: false,
+                    type,
                 },
-                {
-                    new: true,
-                }
-            )
+            }, {
+                new: true,
+            })
 
             await notificationManager(id, 'save_legal_file_event', {
                 type,
@@ -268,7 +258,7 @@ exports.addUserLegalProfile = async (req, res) => {
     }
 }
 
-exports.FindUserLegalProfile = async (req, res) => {
+exports.FindUserLegalProfile = async(req, res) => {
     try {
         const _id = req.params.id
         gfsUserLegal.files.findOne({ _id: ObjectId(_id) }, (err, file) => {
@@ -297,7 +287,7 @@ exports.FindUserLegalProfile = async (req, res) => {
         )
     }
 }
-exports.deleteTwitterChannels = async (req, res) => {
+exports.deleteTwitterChannels = async(req, res) => {
     try {
         const UserId = req.user._id
         const result = await TwitterProfile.deleteMany({ UserId })
@@ -315,12 +305,12 @@ exports.deleteTwitterChannels = async (req, res) => {
     }
 }
 
-exports.deleteTwitterChannel = async (req, res) => {
+exports.deleteTwitterChannel = async(req, res) => {
     try {
         let UserId = req.user._id
         let _id = req.params.id
         let twitterProfile = await TwitterProfile.findOne({ _id })
-        if (twitterProfile?.UserId !== UserId)
+        if (twitterProfile ? .UserId !== UserId)
             return makeResponseError(res, 401, 'unauthorized')
         else {
             await TwitterProfile.deleteOne({ UserId })
@@ -335,7 +325,7 @@ exports.deleteTwitterChannel = async (req, res) => {
     }
 }
 
-exports.deleteGoogleChannels = async (req, res) => {
+exports.deleteGoogleChannels = async(req, res) => {
     try {
         const UserId = req.user._id
         const result = await GoogleProfile.deleteMany({ UserId })
@@ -353,12 +343,12 @@ exports.deleteGoogleChannels = async (req, res) => {
     }
 }
 
-exports.deleteGoogleChannel = async (req, res) => {
+exports.deleteGoogleChannel = async(req, res) => {
     try {
         let UserId = req.user._id
         let _id = req.params.id
         let googleProfile = await GoogleProfile.findOne({ _id })
-        if (googleProfile?.UserId !== UserId)
+        if (googleProfile ? .UserId !== UserId)
             return makeResponseError(res, 401, 'unauthorized')
         else {
             await GoogleProfile.deleteOne({ _id })
@@ -373,7 +363,7 @@ exports.deleteGoogleChannel = async (req, res) => {
     }
 }
 
-exports.deleteFacebookChannels = async (req, res) => {
+exports.deleteFacebookChannels = async(req, res) => {
     try {
         const UserId = req.user._id
         const result = await FbPage.deleteMany({ UserId })
@@ -391,12 +381,12 @@ exports.deleteFacebookChannels = async (req, res) => {
     }
 }
 
-exports.deleteFacebookChannel = async (req, res) => {
+exports.deleteFacebookChannel = async(req, res) => {
     try {
         let UserId = req.user._id
         let _id = req.params.id
         let facebookProfile = await FbPage.findOne({ _id })
-        if (facebookProfile?.UserId !== UserId)
+        if (facebookProfile ? .UserId !== UserId)
             return makeResponseError(res, 401, 'unauthorized')
         else {
             await FbPage.deleteOne({ _id })
@@ -410,13 +400,10 @@ exports.deleteFacebookChannel = async (req, res) => {
         )
     }
 }
-exports.deleteLinkedinChannels = async (req, res) => {
+exports.deleteLinkedinChannels = async(req, res) => {
     try {
         const userId = req.user._id
-        const result = await LinkedinProfile.deleteMany(
-            { userId },
-            { $set: { pages: [] } }
-        )
+        const result = await LinkedinProfile.deleteMany({ userId }, { $set: { pages: [] } })
         if (result.deletedCount === 0) {
             return makeResponseError(res, 204, 'No channel found')
         } else {
@@ -431,7 +418,7 @@ exports.deleteLinkedinChannels = async (req, res) => {
     }
 }
 
-exports.deleteLinkedinChannel = async (req, res) => {
+exports.deleteLinkedinChannel = async(req, res) => {
     try {
         let userId = req.user._id
         let organization = req.params.organization
@@ -439,13 +426,10 @@ exports.deleteLinkedinChannel = async (req, res) => {
             { $unwind: { path: '$pages' } },
             { $match: { 'pages.organization': organization } },
         ])
-        if (linkedinProfile[0]?.userId !== userId)
+        if (linkedinProfile[0] ? .userId !== userId)
             return makeResponseError(res, 401, 'unauthorized')
         else {
-            await LinkedinProfile.updateOne(
-                { userId },
-                { $pull: { pages: { organization } } }
-            )
+            await LinkedinProfile.updateOne({ userId }, { $pull: { pages: { organization } } })
             return makeResponseData(res, 200, 'deleted successfully')
         }
     } catch (err) {
@@ -456,7 +440,7 @@ exports.deleteLinkedinChannel = async (req, res) => {
         )
     }
 }
-exports.UserInterstes = async (req, res) => {
+exports.UserInterstes = async(req, res) => {
     try {
         const userId = req.user._id
         let allInterests = []
@@ -483,19 +467,15 @@ exports.UserInterstes = async (req, res) => {
     }
 }
 
-exports.UpdateIntersts = async (req, res) => {
+exports.UpdateIntersts = async(req, res) => {
     try {
         let id = req.user._id
         let userInterests = req.body.interests
 
-        const interests = await Interests.findOneAndUpdate(
-            { userId: id },
-            { $set: { interests: userInterests } },
-            {
-                new: true,
-                upsert: true,
-            }
-        )
+        const interests = await Interests.findOneAndUpdate({ userId: id }, { $set: { interests: userInterests } }, {
+            new: true,
+            upsert: true,
+        })
         if (interests.nModified === 0) {
             return makeResponseError(res, 400, 'updated failed')
         }
@@ -509,7 +489,7 @@ exports.UpdateIntersts = async (req, res) => {
     }
 }
 
-exports.socialAccounts = async (req, res) => {
+exports.socialAccounts = async(req, res) => {
     try {
         let UserId = req.user._id
         let networks = {}
@@ -517,15 +497,18 @@ exports.socialAccounts = async (req, res) => {
         let channelsTwitter = await TwitterProfile.find({ UserId })
         let channelsFacebook = await FbPage.find({ UserId })
         let channelsLinkedin = await LinkedinProfile.findOne({ userId: UserId })
+        let channelsTiktok = await TikTokProfile.findOne({ userId: UserId })
+
         networks.google = channelsGoogle
         networks.twitter = channelsTwitter
         networks.facebook = channelsFacebook
-        networks.linkedin = channelsLinkedin?.pages || []
-        if (
-            !channelsGoogle?.length &&
-            !channelsLinkedin?.length &&
-            !channelsTwitter?.length &&
-            !channelsFacebook?.length
+        networks.linkedin = channelsLinkedin ? .pages || []
+        networks.tikTok = channelsTiktok ? .pages || []
+        if (!channelsGoogle ? .length &&
+            !channelsLinkedin ? .length &&
+            !channelsTwitter ? .length &&
+            !channelsFacebook ? .length &&
+            !channelsTiktok ? .length
         ) {
             return makeResponseError(res, 204, 'No channel found')
         }
@@ -539,13 +522,10 @@ exports.socialAccounts = async (req, res) => {
     }
 }
 
-module.exports.checkOnBoarding = async (req, res) => {
+module.exports.checkOnBoarding = async(req, res) => {
     try {
         const _id = req.user._id
-        const userUpdated = User.updateOne(
-            { _id },
-            { $set: { onBoarding: true } }
-        )
+        const userUpdated = User.updateOne({ _id }, { $set: { onBoarding: true } })
         const firstNotif = notificationManager(_id, 'buy_some_gas', {
             action: 'buy_some_gas',
         })
@@ -571,7 +551,7 @@ module.exports.checkOnBoarding = async (req, res) => {
     }
 }
 
-module.exports.requestMoney = async (req, res) => {
+module.exports.requestMoney = async(req, res) => {
     try {
         let lang = req.body.lang || 'en'
         var message = req.body.message
@@ -619,7 +599,7 @@ module.exports.requestMoney = async (req, res) => {
     }
 }
 
-exports.support = async (req, res) => {
+exports.support = async(req, res) => {
     const validateEmail = /\S+@\S+\.\S+/
 
     try {
@@ -646,7 +626,7 @@ exports.support = async (req, res) => {
     }
 }
 
-module.exports.notificationUpdate = async (req, res) => {
+module.exports.notificationUpdate = async(req, res) => {
     let _id = req.params.id
 
     if (_id === '{_id}' || !_id) {
@@ -670,13 +650,10 @@ module.exports.notificationUpdate = async (req, res) => {
     }
 }
 
-module.exports.changeNotificationsStatus = async (req, res) => {
+module.exports.changeNotificationsStatus = async(req, res) => {
     try {
         const idNode = '0' + req.user._id
-        const result = await Notification.updateMany(
-            { idNode, isSeen: false },
-            { $set: { isSeen: true } }
-        )
+        const result = await Notification.updateMany({ idNode, isSeen: false }, { $set: { isSeen: true } })
 
         if (result.nModified === 0) {
             return makeResponseError(res, 204, 'No notifications found')
@@ -691,7 +668,7 @@ module.exports.changeNotificationsStatus = async (req, res) => {
     }
 }
 
-module.exports.getNotifications = async (req, res) => {
+module.exports.getNotifications = async(req, res) => {
     try {
         const idNode = '0' + req.user._id
         const arrayNotifications = await Notification.find({ idNode }).sort({
@@ -739,7 +716,7 @@ module.exports.getNotifications = async (req, res) => {
     }
 }
 
-module.exports.changeEmail = async (req, res) => {
+module.exports.changeEmail = async(req, res) => {
     var pass = req.body.pass
     var email = req.body.email
     var user = req.user
@@ -758,10 +735,7 @@ module.exports.changeEmail = async (req, res) => {
             newEmail.expiring = Date.now() + 3600 * 20
             newEmail.code = code
 
-            const result = await User.updateOne(
-                { _id: Long.fromNumber(req.user._id) },
-                { $set: { newEmail } }
-            )
+            const result = await User.updateOne({ _id: Long.fromNumber(req.user._id) }, { $set: { newEmail } })
 
             let requestDate = manageTime()
             let ip =
@@ -792,7 +766,7 @@ module.exports.changeEmail = async (req, res) => {
         )
     }
 }
-module.exports.confrimChangeMail = async (req, res) => {
+module.exports.confrimChangeMail = async(req, res) => {
     try {
         var id = req.user._id
         var code = req.body.code
@@ -806,10 +780,7 @@ module.exports.confrimChangeMail = async (req, res) => {
             return makeResponseError(res, 406, 'code incorrect')
         } else {
             var newEmail = user.newEmail.email
-            await User.updateOne(
-                { _id: Long.fromNumber(id) },
-                { $set: { email: newEmail } }
-            )
+            await User.updateOne({ _id: Long.fromNumber(id) }, { $set: { email: newEmail } })
             return makeResponseData(res, 200, 'email changed')
         }
     } catch (err) {
@@ -821,7 +792,7 @@ module.exports.confrimChangeMail = async (req, res) => {
     }
 }
 
-module.exports.verifyLink = async (req, response) => {
+module.exports.verifyLink = async(req, response) => {
     try {
         var userId = req.user._id
         var typeSN = req.params.typeSN
@@ -870,10 +841,7 @@ module.exports.verifyLink = async (req, response) => {
                     }
 
                     var result = await rp(options)
-                    await GoogleProfile.updateOne(
-                        { UserId: userId },
-                        { $set: { accessToken: result.access_token } }
-                    )
+                    await GoogleProfile.updateOne({ UserId: userId }, { $set: { accessToken: result.access_token } })
                     linked = true
                     res = await verifyYoutube(userId, idPost)
                     if (res && res.deactivate === true) deactivate = true
@@ -939,7 +907,7 @@ module.exports.verifyLink = async (req, response) => {
     }
 }
 
-module.exports.ShareByActivity = async (req, res) => {
+module.exports.ShareByActivity = async(req, res) => {
     try {
         let userId = req.user._id
         let activityURN = req.params.activity
